Moved Simulation & Montecarlo Logging Frameworks to `tardis/io/logger` folder (#1684)
* Moved Logging Frameworks to tardis/io/logger folder

* Fixing some errors with the tests

* Changed simulation_logger.py to logger.py

* Changed implementation to target only TARDIS loggers
Removed list_of_filters to check the filters from the logger itself

* Removed list creation in logging_state & Moved in tardis_logger()
